Add histogram collection script
Plotting is foreseen to take place in two steps:
1) Collect histograms and store them into a root file, such that all information
   is stored in the TDirectory structure of the root file
2) Plot histograms using a second script that retrieves all necessary plots from
   this root file and puts them together onto one plot

This commit contains the script for the first step, that processes one input tuple
at a time. A bash script for automating this over all available input files is also
provided with this commit. However, this is heavily tailored towards the current
use case and will probably undergo some more changes.

Other than that two helper functions have been added to the utils:
1) Checking if a branch is available in a TTree
2) Getting the full trigger path from a uniquely identifying substring