Fix bug in generation where photon was not properly transformed in the same way as the jpsi
This lead to bogus smeared mass distributions since the angular correlation between
the jpsi and the photon was lost.

Smearing is now somewhat tuned, so that it almost matches data distributions but not yet
perfect. Maybe retune, after we see that it has an effect.

Added possibility to generate chic and jpsi data mass distributions directly instead of
trying to obtain them by smearing.
Current settings generate the chic with its natural width and the jpsi as delta.

Added calculation and storage of smeared angles