Move configurable and fixed parts into some helper structs to remove some of the global state
Still some global state present, but this status should make it easier to use a configuration
struct that is now taken as an argument for altering what is generated.

Still can be used standalone in the ROOT prompt, since the config is used as default argument
so that altering makes it possible to steer what is generated.