Fix bug where momentum components could flip sign if deltaP was too large
Especially for the photon smearing it was possible that deltaP was negative with
an absolute value larger than the particle momentum leading to a sign flip in
the smeared momentum components, effectively rendering the photon a random one.
Using the absolute value fix for p + deltaP fixes this.

Note: This is only a viable solution because the fraction of events for which
this happened was small. If this changes it is possible that the smearing does
no longer follow the input distribution, since the actual smearing might cap
the tails of the distributions