Add possibility to get uncertainty bands of the photon efficiencies
Since I couldn't find any straight forward way of getting a TF1 evaluated at different
parameter values taking into account correlations between the parameters this commit
comprises a rather elaborate workaround for which I tried to get at least a part of the
work as general as possible so that it might be reused.
The main problem is that ROOT only enables the retrieval of uncertainty bands if the
fit result is still available in its entirety (at least I didn't find another way). Since
I also didn't manage to "construct" such a fit result from the available information:
- the function
- the parameter values
- the complete covariance matrix
I concluded that the "easisest" way to go forward is to calculate the uncertainty as a
function of pT and then use this function to calculate the uncertainty band. I didn't
go all the way through to store this result as a TF1 (though I think it should technically
be possible) but instead decided to determine the value of the function at several values
of pT and store this as a TGraph.

The uncertainty function is "simply" calculated using error propagation taking into
account the correlations between the free parameters. Since the parametrization is
differentiable as much as possible is done symbolically and only at the very end actual
values for the parameters are introduced.
In this way the efficiency and its uncertainty band are created as a TGraphErrors by
evaluating the function (using the central values of the fitted parameters) and the
uncertainty function (using the correlation coefficiencts and the uncertainties of the
fitted parameters) at different values of pT.
In the end a TGraph that either has the y-values at the upper or lower uncertainty band is
stored and can be used in the same fashion as the TF1 that is stored if no uncertainties
are desired (not passing the --sigma parameter or passing 0 to it).