Fix minor bug in range frinding from fit.
Accidentally defined the stepsize as negative, which lead to the second initial value
to lie below the first initial value (so that they no longer enclose the root). However,
since it is not a precondition of the secant method to have the starting points enclosing
the root, this bug was inconsequential.