Add general functions to get extremal x and y values from ROOT plotables.
Previously this functionality was only implemented for TH1s. Now it is also
available for TGraphs and TF1s. Since python doesn't allow overloading, but
ROOT does not provide a common interface for this, the top-level functions in
plot_helpers are just wrappers that call the corresponding functions for all
elements