Add fit model that takes a json config file to define the model
To be more flexible and have the model actually be part of the configuration of the fit
added a class inheriting from FitModel that takes a json config file (similar as chib
already does). Added the configuration for the chic model with the same settings as it
is as of this commit in the ChicMassModel.

TODO:
- Switch all states to the config file approach and remove the individual models and
  cleanup the fitting code
- Implement a way so that the variables that should be fixed after a costh integrated
  fit are part of the configuration (i.e. not just event yield variables as it is at
  the moment