Change pairing of colors and markers in default_attributes
Previously if unique_colors=False was used the colors were exhausted first, before
switching to a new marker style. Now we iterate over both in pairs so that the first
elements are now exacatly the same as it was before with unique_colors=True, except
that we do not stop when we run out of marker styles.
This new pairing ensures easier to discern combinations up to higher number of plots
than any of the previously possible combinations.
The unique_colors kwarg has been marked with a deprecation warning and will be ignored.