Enable interactive button naming as soon as it is available
Before ipywidgets version 7.2.0 the buttons displayed the function name they
were bound to, which was not perfect, but at least allowed to distinguish them.
With the new feature, this is no longer true, but the buttons can be named,
so that in the end with some work this is actually nicer.