Added a different sampling kernel ot avoid extreme oversampling of the far edges
Using a Gaussian will either lead to a too "democratic" sampling or to an extreme
oversampling of the very edges. Added a sampling kernel that samples the costh space
in a stepwise fashion starting at low values in the middle and then increasing to
higher values in two steps. The boundaries of the steps are made pt dependent to
account for the different widths of the reconstructed distributions