Make first steps towards a more flexible histogram storage possibility
Under the hood everything uses ROOT::THn now, which should allow to decide the
dimensionality at runtime in the end. Currently it is done at compile time and
also the binning and the used varialbes are fixed / hardcoded at the moment.

All the configuration / initilization should now be done in one place and the
used variables as well as the functions that calculate the weights, etc. must
be specified via a lambda function to pick the corresponding variables from
the chicpolgen scope.