Enable Sumw2 for all projections before returning from project
For some reason after projection of a THn its status regarding the error calculation
is not the same as it was before. Especially if a histogram returned True for
GetCalculateErrors() before the projection, the resulting histogram returns False for
the same call. Simply setting Sumw2 regardless of the status solves all my problems
since that should be the default in any case.