Pull generation of pT and rapidity into generate-check-reject loop of angular variables
When using importance sampling and not regenerating pT and rapidity for each event (as
it was before) introduces a bias into the generated pT / M distribution (was checked with
the J/psi). Pulling it into the loop that deals with the angular distribution forces the
values to be newly generated for every event and not reusing "old" values from a previous
event that was rejected (due to importance sampling).