Implement weighted mass fits.
Corrections will be applied via weights so using these weights has to be possible in
the mass fits. Majority of changes is related to this and is mainly related to having
the weights available when necessary and actually using them.

Unrelated to this now the fit results and the covQual of the fit are now also stored
into the workspace so that they can be retreived from there instead of from logfiles.