Make it possible to use already defined variables for event yields and return success of define_model
Previously the importing was done with the expectation that the event_yield property
was defining a variable in the workspace. This makes it impossible to fix the yield
to another variable.
In principle this was already possible before, but this would give errors during the
import (although it seemingly still worked). Now it is done properly and the definition
of the model fails if the event yield is not already defined or not defined directly.