Generalize handling of graph creation in make_fit_plots
Up to now only the free parameters in the fits were produced as a function of costh,
additionally r_chic2_chic1 was manually handled.
Now all the yields (even if they are not technically free parameters in the fits),
r_chic2_chic1 plus all free parameters will always have a graph.
Some smaller changes like introducing a check if a variable is dependent had to be
done, to choose the right strategy for obtaining the uncertainties.