Fix bug in plotting leading to wrong draw_option in some cases
If a plot_hist is used in mkplot then the 'same' option is added to the drawOpt kwarg.
However, in plot_on_canvas this was added a second time unconditonally for all but
the first pltable. This repetition of the 'same' option leads to the 'e' being interpreted
in a wrong way and so to inconsistent plotting options for the different pltables