Make mkplot handle THStack properly
mkplot should now be able to properly handle THStack objects by giving each of
the stacked histograms appropriate attributes and also considering them in
assigning legend entries.

There is a bug in ROOT up to v6.16/0 that breaks things when calling GetMinimum
or GetMaximum on a THStack object. A run-time warning for this is in place.