Automatically unpack get_var if it is still a tuple in _get_var
_get_var expects the variable and the np_func as two separate arguments, but
parse_func_var returns a tuple, so that calling _get_var directly with the
output of that is not possible.

Automatically unpacking the tuple if it has length 2, avoids having to introduce
some more handling code on the calling site, with the downside that it is now
easier to break this function when calling it.

Tests still pass, although that is literally no guarantee in this case, because
they simply do not really test this scenario