Print global chi2 / ndf on all fit result plots
For simultaneous binned fits the chi2 / ndf is not something that can be
interpreted for all bins individually. Instead only the "global" chi2 / ndf has
statistical meaning.

In order to make this happen the signature of get_chi2_ndf was slightly changed,
but all calls in the repository should be updated accordingly with this commit