Switch to symmetric uncertainties if asymmetric ones are not there
MINOS reports status 0, as soon as the asymmetric uncertainties for one of the
parameters of the fit can be calculated. This does not imply that all asymmetric
uncertainties are properly calculated. In such a case the output graph will now
switch to symmetric uncertainties and emit a warning.

NOTE: For now the whole graph uses symmetric uncertainties in such a case,
although it is also possible to make this decision on a point-by-point (i.e.
fit-by-fit) basis in principle.