Fix calculation of BIC and AIC to use the actual min nll
When Offset(True) is used, the value returned by RooFitResult.minNll() is the
result of the minimization, which is not the value of the negative log
likelihood at the minimum. Storing the RooNLLVar and evaluating it at the
minimum actually evaluates it at the currently set parameter values, which will
give the minimum even if an offset is used in the minimization for numerical
stability.

To avoid having to rerun all the fits up until now it is created when plotting
if it is not already stored in the workspace.