Fix bug and return to nominal model
Previously Nchic0 was defined slightly wrong, instead of

Nchic1 * r / (1 + dlth * costh**2)

it read

Nchic1 * r * (1 + dlth * costh**2)

which is not the correct relation that can be obtained by solving Nchic1 /
Nchic0 for Nchic0 (and using that chic0 is unpolarized). However, in the end
this leads to only very small changes, since the chic0 peak is too small to have
a real influence on the chic2 / chic1 ratio.

Also defined the background model again, to have constant sigma, linear mu and
free lambda.