Update calc_ppd.py to be able to handle the directions independently
With a previous commit the two directions (costh and phi) were coupled too
tightly. Since studies have shown that the results mainly depend on the priors
but not so much on whether one or both ratios are used, this commit reverses
parts of that earlier commit, while in principle keeping all the functionality.

To still be able to have a quick overview on how to most easily interpret the
produced histograms, a README was added.

The scripts for submitting batch jobs have alse been updated, but are not yet
tested (!)