Use symmetric uncertainties for ppd comparison plots
In principle the uncertainties determined from the quantiles of the ppd are
asymmetric, but that leads to weirdly asymmetric correlated uncertainties. This
is an attempt to mitigate this by calculating the correlated uncertainties using
the mean value and the rms of the ppds.

Leaving the old code in place in case it has to be (temporarily) reversed