Add possibility to have "publication quality" mass fit result plots
This has really been shoehorned in at this moment. The ideal solution is to
get (almost) rid of the non-publication quality versions of the plots at some
point.

NOTE: The legend position for most fit configurations has to be updated to be on
the left side of the chic peaks