Make it possible to have mixed hist settings in hist2d and hist3d
ROOT does not offer a TH3D constructor that has mixed axis definitions, where it
is e.g. possible to define a number of bins and a min/max value for the x-axis
but a binning with defined bin-borders along z. Converting all the settings to
the same format makes this possible.