Add possibility to fix lambda1 and only fit Delta lambda
This yields basically the same results as a simple chi2 fit where lambda1 is
fixed to a given value. However, it allows the processing of the output within
the framework without having to deal with another format of the fit results.

If lambda1 is fixed the prior for lambda2 is "opened" to reach further than the
physically allowed range to not restrict the results of Delta lambda
artificially. Without this the correspondence to a simple chi2 fit would be
lost.